Move orm related code to its own module. See #6
